histogram wip

+++ src/salon_core/src/engine/ops/histogram.rs
@@ -36,7 +36,7 @@ impl ComputeHistogramImpl {
         );
 
         let buffer_props = BufferProperties {
-            size: 4 * 256 * size_of::<f32>()
+            size: 4 * 256 * size_of::<u32>()
         };
 
         let output_buffer = value_store.ensure_value_at_id_is_buffer_of_properties(
@@ -48,7 +48,7 @@ impl ComputeHistogramImpl {
         self.runtime.queue.write_buffer(
             &output_buffer.buffer,
             0,
-            bytemuck::cast_slice(&[0.0; 4 * 256]),
+            bytemuck::cast_slice(&[0u32; 4 * 256]),
         );
 
         let bind_group = self
